move getUser fetch call to the profile page
