+++ components/Statistics/index.js
@@ -1,33 +1,13 @@
-import { useState, useEffect } from "react";
-
-import { useAuth } from "../../context/AuthContext";
-import UserModel from "../../model/user";
 import StatsCard from "../StatsCard";
 import checkHighestColor from "../../utils/checkHighestColor";
 import checkLowestColor from "../../utils/checkLowestColor";
 import pickLowIcon from "../../utils/pickLowIcon";
 import pickHighIcon from "../../utils/pickHighIcon";
 import SAMPLE_FOOD_DATA from "../../sampleData/userSampleFoodData";
-import Loading from "../Loading";
 
 import styles from "./styles.module.css";
 
-const Statistics = () => {
-  const [userObj, setUserObj] = useState();
-  const [loading, setLoading] = useState(true);
-  const {
-    authedUser: { uid },
-  } = useAuth();
-
-  useEffect(() => {
-    async function fetchData() {
-      const userResponse = await UserModel.getUser(uid);
-      setUserObj(userResponse);
-      setLoading(false);
-    }
-    fetchData();
-  }, []);
-
+const Statistics = ({ userObj }) => {
   const lowestColor = checkLowestColor(SAMPLE_FOOD_DATA);
   const highestColor = checkHighestColor(SAMPLE_FOOD_DATA);
 
@@ -40,9 +20,6 @@ const Statistics = () => {
   const lowestColorData = capatalizeString(lowestColor.color);
   const highestColorData = capatalizeString(highestColor.color);
 
-  if (loading) {
-    return <Loading />;
-  }
   return (
     <div className={styles.container}>
       <p>Statistics</p>
